fix OverAll UI
UI now automatically updates based on the data from the database.
When the user logs in, their username is saved; upon logout, the username is deleted. This allows users to remain logged in between game sessions.
Implemented scene management improvements to prevent freezing when returning to the game after quitting, ensuring a smooth user experience.